+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ### 2025-06-16
+- [Patch v6.9.55] Adjust menu full_pipeline and return code
+- New/Updated unit tests added for tests/test_projectp_script.py
+- QA: pytest -q passed (374 tests)
+
+### 2025-06-16
 - [Patch v6.9.54] Improve fast test mode in run_tests
 - New/Updated unit tests added for tests/test_run_tests.py
 - QA: pytest -q passed (9 tests)